Bug 1858082 - Wrap the drag service getter into a try block r=whimboo,webdriver-reviewers
The drag service is never instantiated if it's in the headless mode.
And in this case, `NS_ERROR_XPC_GS_RETURNED_FAILURE` exception is thrown.
Therefore, we need to wrap the service getter with a `try` block and
return `null` if it fails.

Differential Revision: https://phabricator.services.mozilla.com/D190580

UltraBlame original commit: 25aeae00c1018f195b74fa9458f71deb806796e2