Bug 1857903 - fix VPN Promo check to read the correct pref, r=omc-reviewers,aminomancer
There was a typo in the preference name used for the primary list of all regions where VPN promos were allowed to be shown.  As a result, VPN promos were shown to regions from a built-in fallback list, which was smaller.  This patch fixes the typo and adds a unit test to ensure the code gets its regions from the correct pref.

Differential Revision: https://phabricator.services.mozilla.com/D190628

UltraBlame original commit: 9bf71a837fe38f4f6a859419607447df9711866f