Bug 1782962 - Add telemetry for dom.popup_allowed_events pref r=edgar,chutten
It can be manually added to control what events allow popups.

If the popup blocker is rewritten to use the new user activation model
and this is relied upon it would be a blocker (needed to be implemented
in the new model).

Differential Revision: https://phabricator.services.mozilla.com/D190442

UltraBlame original commit: 9ba44fb4eeafba4229fe28a6f3b42c79d8d05692