Bug 1858110 - Make a rule easier to override. r=pbz,desktop-theme-reviewers,dao
This makes the rule in [1] properly override it, which is enough to fix
this bug.

Ideally the view would be able to shrink, but we'd have to figure out
how to not regress bug 1822819 for that, which I don't have time for
right now.

[1]: https://searchfox.org/mozilla-central/rev/81f368dab93fff035ce7fcc376e16990e89dd5ec/browser/themes/shared/customizableui/panelUI-shared.css#1989

Depends on D191054

Differential Revision: https://phabricator.services.mozilla.com/D191055

UltraBlame original commit: b1dce4bb43dfffcc62fed8bdfb0dc23a9ff6a1de