Bug 1852237 - Remove macOS sheet attachment hack. r=desktop-theme-reviewers,dao
This was needed so that stuff like the Cmd+Q dialog was positioned
correctly, because such dialog used a native macOS sheet.

But that's no longer the case, so I don't _think_ we need this. If we
needed this, we could use -moz-default-appearance much like we do on
Windows[1]. But that doesn't seem necessary right now.

[1]: https://searchfox.org/mozilla-central/rev/81f368dab93fff035ce7fcc376e16990e89dd5ec/layout/painting/nsDisplayList.cpp#3122

Differential Revision: https://phabricator.services.mozilla.com/D191051

UltraBlame original commit: 7c819a13e238e8c7d73d1016561bd5241152ad1e