Bug 1859265 - Update cached metadata before updating add-ons. r=rpl,amejiamarmol,geckoview-reviewers
We currently do not refresh the cached metadata in the add-ons
repository because the update timer isn't enabled on Android.

This patch calls the add-ons repository update logic to refresh
the cached metadata when an add-on is checked for updates. This
is similar to what we're doing when we perform a (full) background
update check on Desktop.

Differential Revision: https://phabricator.services.mozilla.com/D191081

UltraBlame original commit: 36e5163d2528daaa5669d0f9284470b070c8882e