Bug 1857486 - Adjust geckoview.logging value based on debug build. r=geckoview-reviewers,owlish
Currently geckoview-prefs.js attempts to set geckoview.logging to
"Debug" in nightly bulds and "Warn" in release and beta builds. This
logging is rather verbose, so we should disable it in optimized
nightly builds. ie set the default value based on whether it is a
debug build rather than the release channel.

However, the value in geckoview-prefs.js is currently ignored
altogether in favour of the value set in GeckoRuntimeSettings.java.
This patch therefore removes the pref from geckoview-prefs.js to avoid
confusion, and updates the default in GeckoRuntimeSettings based on
whether it is a debug build.

Differential Revision: https://phabricator.services.mozilla.com/D190280

UltraBlame original commit: a1b07881519bf27f1f641d395e2edcf4ee4fd3a3