Bug 1858842 - Fix nsMainThreadPtrHolder's destructor during shutdown. r=xpcom-reviewers,mccr8
As it is written, if the MOZ_ASSERT(mMainThreadEventTarget) fails, we'd
release the object on the current thread, which is not what we want to
do.

NS_ReleaseOnMainThread deals with shutdown leaking the object, which is
safer. Nobody overrides mMainThreadEventTarget so make the code simpler
while at it.

Differential Revision: https://phabricator.services.mozilla.com/D191108

UltraBlame original commit: 57897c091b1fa9d30f7141baa838c2ae4bd65cfb