Bug 1858807 - Track individual rust crate build times. r=firefox-build-system-reviewers,sergesanspaille
As of now, cargo only supports HTML timings output. There is an unstable
flag for a json output, but we can't use that because it requires a
nightly cargo. So we do our best effort to get the data we need off the
HTML, but since the exact details of the format of that file are not
set in stone, we also prevent any error from being unhandled. It's
better not to have the data than break the build.

Differential Revision: https://phabricator.services.mozilla.com/D190884

UltraBlame original commit: 45208ee996d78e01fcb532ae9830399954c4f5fe