Bug 1860051 - Handle TestRoot in manifestupdate, r=Sasha
Upstream PR 42426 replaced the dict of paths with a class. To enable that PR to land
in m-c update the code to handle either a dict or a class for now. Once the PR is landed
we can allow it to simply handle the class.

Differential Revision: https://phabricator.services.mozilla.com/D191625

UltraBlame original commit: 186e5ca1f7a4dc4b4bded55549979f2cf6902c27