Bug 1860766 - Clean-up in-content menulist styles. r=dao,desktop-theme-reviewers
Before bug 1860385, we used the 16px image (because the list-style-image
was set on the icon, and in-content/common-shared.css rule didn't
override it). But the dropmarker was still 12px wide, and the 4px
margin-inline-end of common-shared.css did fix it.

That's pretty messy. Remove the dropmarker sizes so that the dropmarker
takes the right size based on the dropmarker icon. That means that if
somebody wants a 16px icon they can just override list-style-image on
the dropmarker.

Use the 12px dropmarker arrow since that's what we also do for <select>,
and seems to be the general intention.

This also avoids the need of manually positioning it on windows to be
centered-ish vertically.

Differential Revision: https://phabricator.services.mozilla.com/D191736

UltraBlame original commit: 564a91732aa14455a09ac958b22e1b0bde7c6a63