Bug 1860954 don't store the graph lookup key in the graphs hashtable r=padenot,pehrsons
The necessary values for the lookup are now all in the graph itself.

I suspect this would also be possible with nsTHashtable but HashSet is
designed for a lookup class that differs from the stored class and requires
less boilerplate.

Depends on D191831

Differential Revision: https://phabricator.services.mozilla.com/D191832

UltraBlame original commit: a1d1cccb0f7fddf117ebdd1221d329be9bd03c56