Bug 1857687 - Disable thread safety analysis on a few problematic functions. r=nika,necko-reviewers,valentin
The functions return a reference to a value that is normally protected
by a lock in an unlocked manner, but they do so on purpose.

One of them can be annotated with the lock that is being held when it's
called, though, so we do that.

Differential Revision: https://phabricator.services.mozilla.com/D191707

UltraBlame original commit: 46e5245db90d62939d82dd52ef22c509f586a900