Bug 1861288 - Improve coverage of the pre-export and export phases in profiles. r=firefox-build-system-reviewers,ahochheiden
At the top-level, what we currently have is something akin to:
  export: some deps
    echo BUILDSTATUS TIER_START export
    make recurse_export
    ...

In practice, what this means is that `some` and `deps` are built outside
the export tier/phase.

Since recurse_export is itself a recipe with dependencies, we can "just"
move `some` and `deps` to become dependencies of recurse_export instead.
But since in non top-level directories, recurse_export is not a thing,
we still need to keep the dependencies on export there.

Unfortunately, we have one instance of something built during
recurse_export that relied on its undeclared/undeclarable dependencies
being built before recurse_export, so we need to accomodate for that.

Also, for some reason, recurse_pre-export was a two-colons recipe, while
the other recurse_* recipes were single-colon, so fix that.

Come to think of it, the same probably applies to misc and libs, which
this would mechanically fix alongside.

Incidentally, this also makes those things that ran before recurse_* run
in parallel of the other things that run during recurse_*.

Differential Revision: https://phabricator.services.mozilla.com/D191913

UltraBlame original commit: a8fd3a9117b2f9e1409074b593e2bdbd0761a1d8