Bug 1843335 - [devtools] Add border to matching filter element in rules view. r=devtools-reviewers,ochameau,accessibility-frontend-reviewers,ayeddi.
Since it would be tricky to get a color that would give us a good contrast,
we add a bottom border to matching elements to differentiate them from the
other ones.
This also matches the style that we have in Debugger search results.

We're using an inset box-shadow so we don't impact the layout.

This patch also modifies to which elements we apply the "matching" style for
properties. Previously we were applying to individual elements on the property
container, which would have been hard to make a seamless border to work.
Instead, we're putting the styling on the whole property itself.
As the padding on it + negative margin on `.ruleview-namecontainer` was making
it look off, we're simply removing them.
This was used to ensure that wrapping lines wouldn't be placed directly under
the checkbox (see https://bugzilla.mozilla.org/show_bug.cgi?id=1436185#c8 ).
We can replicate this with a hanging text-indent instead.

Differential Revision: https://phabricator.services.mozilla.com/D190691

UltraBlame original commit: 6e0f177caa133a24fffc896b82fb3f5dc2879733