Bug 1862087 - Add test labels for two non-userfacing PanelMultiView tests in Customizable UI. r=Jamie,Gijs
There are 2 failing tests for the `PanelMultiView`s that appear to create mock toolbarbutton and panel for the purposes of testing PanelMultiView, but the anchor (`gAnchor`) or a toggle (`gToggle`) do not have an a11y label (because they were not really intended to be seen by a user).

We are adding a test label for each of them to keep the a11y_checks test harness happy and removing the `fail-if` annotations from the test manifest that was added by bug 1854536 before the investigation in the meta bug 1854537.

Differential Revision: https://phabricator.services.mozilla.com/D192289

UltraBlame original commit: 4fc6d2c75998ae6fcea2f75a12c3db5f6d1830b7