Bug 1722322 - Fix two bugs in nsContentSecurityManager::GetSerializedOrigin r=necko-reviewers,dragana
The first bug is according to https://fetch.spec.whatwg.org/#serializing-a-request-origin,
the algorithm should return "null" if the request has a redirected-tainted origin.

The second bug is the comparison of whether this is a cross origin redirect missed
the `not operator`.

Differential Revision: https://phabricator.services.mozilla.com/D151078

UltraBlame original commit: f52b1f28f753170787e6d0cdca92606caee61249