Bug 1722322 - Update image-not-fully-visible.html to make it work on Firefox Android r=emilio
The meta tags for width and height are needed to make sure the
calculation for the image position uses device sizes
rather than virtual sizes.

`minimum-scale` is used to ensure the `overflow:hidden` content
is not scrollable on Firefox Android, so that the LCP size can
be computed correctly.

Differential Revision: https://phabricator.services.mozilla.com/D192586

UltraBlame original commit: 7a75717f127c90bbeb60371ee8ea99e562c97657