Bug 1861296 - Update blocklist telemetry expiry to 132 r=rpl
And remove the mention of the blocklist.mlbf_enabled scalar, since its
implementation was removed in bug 1820155. Note that mlbf_enabled
defaulted to true since Firefox 78 (bug 1631018), and that the pref to
revert it was removed in Firefox 91 (bug 1706391). The latter was
desktop-only, but the telemetry wasn't sent on Android anyway. The
Glean-based telemetry was introduced in Firefox 115 (bug 1820155) along
with the removal of mlbf_enabled.

Differential Revision: https://phabricator.services.mozilla.com/D192743

UltraBlame original commit: 58e4bd7fa9a8acbe85d321d1c4b33b3f9636f28f