Bug 1861998 - Modify cron so that android CaR only runs once a day. r=perftest-reviewers,taskgraph-reviewers,ahal,sparky
This patch modifies the target tasks so that we don't run the sp3 tests
more than once a day for the Android custom-CaR application. This is for
consistency as this is already taken care of in the
`custom-car_perf_testing` which handles custom-car for all platforms.

Differential Revision: https://phabricator.services.mozilla.com/D192518

UltraBlame original commit: 0cb5b12aa2dffce0f3c4119e0e4a1bdeeb5bbd31