Bug 1854010 - Restrict fixed/sticky hit test info to items fixed to the root. r=hiro
Fixed and Sticky position side bits should only be necessary for fixed
or sticky position items that are fixed or stuck to the root content. As
a result, there is no need to add the fixed or sticky position hit test
info to items that are not.

Differential Revision: https://phabricator.services.mozilla.com/D192272

UltraBlame original commit: 767972bb0fc3b9f4780ebc4ed520507248346a84