Bug 1863367 - Part 1. Add Image::BuildSurfaceDescriptorBuffer and use for RDD readback. r=gfx-reviewers,lsalzman
This patch adds a new method for Image subclasses to override in order
to avoid extra copies when the ultimate destination of a readback is
into an IPDL Shmem. It also makes RemoteDecoderManagerParent's
RecvReadback take advantage of it.

Differential Revision: https://phabricator.services.mozilla.com/D192883

UltraBlame original commit: 9b9786fd643840e2204b8eabac144a67c89f0fb7