Bug 1848177 [wpt PR 41410] - sensors: Convert existing Generic Sensor web tests to test_driver, a=testonly
Automatic update from web-platform-tests
sensors: Convert existing Generic Sensor web tests to test_driver (#41410)

All Generic Sensor web tests are now using virtual sensors and their
implementation in WPT's testdriver.js and ChromeDriver/content_shell.
These tests no longer depend on generic_sensor_mocks.js (or Mojo mocks
in general), which has been kept only for the Device Orientation tests
at this point, and should be fully interoperable now. This also
exercises the code in //content and //services.

Compared to the JS mocks-based implementation, one of the biggest
changes is that there is no equivalent to a MockSensor that takes an
array of readings and periodically rotates and reports them to
consumers. This was an implementation detail that was kind of hard to
justify in spec terms, and the interaction between the JS timer and the
C++ code in services could lead to races between shared buffer values
and expected values (bug 1073865 is probably related, for example).

We now follow the new Automation section in the Generic Sensor spec and
explicitly require reading updates to be done via calls to
test_driver.update_virtual_sensor(). These calls might not result in a
change in readings if the sensor is not active or if the
threshold/rounding checks fail, so the tests are careful to check that a
reading has gone through when necessary. Furthermore, given how the code
in //services is implemented it is not possible to assert that a call to
update_virtual_sensor() will resolve _before_ a "reading" event is
emitted, so some Promise.all() calls are required.

The iframe tests have been essentially rewritten for clarity and to test
the right things:
* iframe_sensor_handler.html is used only by the cross-origin test. The
  same-origin one simply creates the sensor in the test file directly.
* send_message_to_iframe() does not take a |reply| argument anymore.
  Instead, it propagate the replies to the callers so they can do
  whatever they need and we do not need to come up with values like
  "success" in replies.
* The cross-origin and same-origin tests follow a workflow that is
  easier to follow and mostly relies on timestamps that are part of the
  public API to determine whether readings are being provided or not
  depending on frame focus.

Now that the //content code is being exercised, the Ambient Light Sensor
and Magnetometer web tests need the GenericSensorExtraClasses feature to
be enabled, which is done by the newly-created
generic-sensor-extra-classes virtual test suite.

Finally, the referenceFrame test in generic-sensor-tests.js has been
disabled and a copy was added to wpt_internal, as there is no
cross-platform way to set an orientation angle and in wpt_internal it
can be set via Internals' setMockScreenOrientation().

Bug: 1278377, 1471996
Change-Id: Ie556c6d7bcfc0b2b84abc5f0770f6b3120ec81a2

Co-authored-by: Raphael Kubo da Costa <raphael.kubo.da.costaintel.com>
--

wpt-commits: ea2a5d683d75e38809bbe9ec5fb238267054b747
wpt-pr: 41410

UltraBlame original commit: ff140fd800cf4c913018e69a1566664de55a1370