Bug 1605894 remove an unused LoopbackTone instance r=padenot
As the comment indicates, this LoopbackTone is unused.  It is an unnecessary
use of the global TEST_AUDIO_FREQ.  If verification of gUM call stream
contents should be required in the future, then the test can choose which
frequency to use and match this with verification.

Having less realtime audio running can simplify debugging.

Depends on D193514

Differential Revision: https://phabricator.services.mozilla.com/D193515

UltraBlame original commit: 77d5eb0444c869e8c148c11266bd305091df369e