Bug 1442018 - Avoid crashing in BCPaintBorderIterator if mCellMap was not found. r=layout-reviewers,emilio
This does not fix the underlying issue, which is that when the table
got fragmented, we failed to maintain the cellmap (attached to the
first-in-flow) to keep track of the additional rowgroup frame created
to go in the overflowing part of the table.

So this is a wallpaper patch just to prevent the crash here (confirmed
with testcase 5); it'll still throw NS_ASSERTIONs in a debug build,
and in the "bad" case (which depends on details of scaling) some of
the borders will be missing from the print preview/output, but that
seems preferable to crashing.

Differential Revision: https://phabricator.services.mozilla.com/D193660

UltraBlame original commit: fa9a9180e2440aa0836173e1aa3cfe27c999f48c