Bug 1864318 - Faster elfhack core loop r=glandium
This is 20% faster on my setup, and according to llvm-mca, the IPC for
the false branch (which is the hottest one) goes from 3 to 5.7, thanks
to unrolling and conditional moves.

Basically, most of the bits of `bits` are generally set, so it's ok to do a
few more extra operations as we do them faster.

Differential Revision: https://phabricator.services.mozilla.com/D193366

UltraBlame original commit: 47c58e2355e9db64bb3ce463cb3f976c2496d4b1