Bug 1865383 - Check slice budget when we start marking or sweeping r=sfink
The patch forces a budget check when we start marking or sweeping unless we
have just started a GC slice in the current state.

This turned out more complicated than expected because the change caused a test
that called gcslice(1) in a loop to not make any progress. This was because the
original stepAndForceCheck() method ate the single unit of work budget. The fix
was to remove the step part of this method. I don't think any of the callers
need it to step the budget as well.

Differential Revision: https://phabricator.services.mozilla.com/D193925

UltraBlame original commit: b4b1c9fc65090f4a10a3d214d98766de25c14a68