Bug 1857279 - Make list of `try` subcommands static to get around a circular dependency r=ahal
If a new subcommand is added to `try` this will need to be manually
updated. The first person to set a `try` default for this theoretical
new `try` subcommand will encounter an error, but from that error it
should be relatively straight forward to add it to this list. I think
with the infrequency that we add new commands, and this being the only
way to solve a different problem, this solution is acceptable.

Differential Revision: https://phabricator.services.mozilla.com/D192373

UltraBlame original commit: d53aa284c76913d5d692911b11af2fa30922ce18