Bug 1799173 - Override Predispatch/PostDispatch of CreateImageBitmapFromBlobRunnable to mute the noise of WorkerRunnable dispatching assertion. r=dom-worker-reviewers,asuth
CreateImageBigmapFromBlob is a behavior across the owning thread and the main thread. It starts from the owning thread, completes the IO on the main thread, and returns to the owning thread in the end.

However, when CreateImageBitmapFromBlob is in a nested worker, it uses CreateImageBitmapFromBlobRunnable to return to the owning thread. And It hits the assertion of WorkerRunnable::PreDispatch(), which supposes the runnable should be dispatched from its parent thread.

According to the design of CreateImageBigmapFromBlob, it should not be restricted by the assertion of WorkerRunnable::PreDispatch(), and it should always be dispatched back to the owning thread. Therefore, This patch overrides PreDispatch() and PostDispatch() of CreateImageBitmapFromBlobRunnable so that it can work correctly.

Differential Revision: https://phabricator.services.mozilla.com/D193553

UltraBlame original commit: acb4edab4daf5cfc34700773250ec5850a4df350