Bug 1862768: move mac attribution cache to a new file r=nalexander
While investigating why initial macOS attribution data had a lower attribution rate than expected I realized that one reason why this may be the case is because we're using the same file as a cache for it. This means that even if a user has paved over with a fresh install, we will pick up the previous attribution data (even if it was null).

While doing this, I think we should also accommodate having separate attribution data per channel, which should keep the data cleaner. Release, Beta, and ESR use `Firefox.app`, which makes the likelyhood of a paveover with a different channel reasonably high.

Differential Revision: https://phabricator.services.mozilla.com/D192628

UltraBlame original commit: f64b5449a78cad22499552563782697f65403e5b